Aufteilung Code-Elemente und Initialisierung Code- / Anwendungsdokumentation
