+++ app_documentation.md
@@ -1,0 +1,219 @@
+# ![Project Logo](./logo.png)
+
+# Project Title
+
+## Application Documentation & Development Documentation
+
+---
+
+## Table of Contents
+
+1. [Cover Page](#cover-page)
+2. [Project Overview](#project-overview)
+3. [Table of Contents](#table-of-contents)
+4. [Application Documentation](#application-documentation)
+    - [Introduction](#introduction)
+    - [Features](#features)
+    - [Installation](#installation)
+    - [Configuration](#configuration)
+    - [Usage](#usage)
+    - [Screenshots](#screenshots)
+    - [Troubleshooting](#troubleshooting)
+    - [FAQ](#faq)
+    - [Contact & Support](#contact--support)
+5. [Development Documentation](#development-documentation)
+    - [Architecture Overview](#architecture-overview)
+    - [Project Structure](#project-structure)
+    - [Development Setup](#development-setup)
+    - [Coding Guidelines](#coding-guidelines)
+    - [Key Modules & Components](#key-modules--components)
+    - [Database Schema](#database-schema)
+    - [API Endpoints](#api-endpoints)
+    - [Testing](#testing)
+    - [Deployment](#deployment)
+    - [Changelog](#changelog)
+    - [Contributing](#contributing)
+    - [License](#license)
+
+---
+
+## Cover Page
+
+|                |                                             |
+|----------------|---------------------------------------------|
+| **Project:**   | [Project Name]                              |
+| **Version:**   | 1.0.0                                       |
+| **Date:**      | YYYY-MM-DD                                  |
+| **Authors:**   | [Your Name], [Team Members]                 |
+| **Contact:**   | [your.email@example.com]                    |
+
+---
+
+## Project Overview
+
+Briefly describe the purpose and goals of your project.
+
+---
+
+## Application Documentation
+
+### Introduction
+
+Describe the application, its purpose, and the target audience.
+
+---
+
+### Features
+
+- Feature 1
+- Feature 2
+- Feature 3
+
+---
+
+### Installation
+
+```bash
+# Step-by-step installation instructions
+git clone https://github.com/youruser/yourrepo.git
+cd yourrepo
+pip install -r requirements.txt
+```
+
+---
+
+### Configuration
+
+Explain configuration options, environment variables, or files.
+
+---
+
+### Usage
+
+Provide usage examples, command-line arguments, or UI navigation.
+
+---
+
+### Screenshots
+
+Include relevant screenshots or GIFs.
+
+![Screenshot](./screenshots/screen1.png)
+
+---
+
+### Troubleshooting
+
+| Problem            | Solution                |
+|--------------------|------------------------|
+| Common Issue #1    | How to fix it          |
+| Common Issue #2    | How to fix it          |
+
+---
+
+### FAQ
+
+**Q:** Frequently asked question?  
+**A:** Answer.
+
+---
+
+### Contact & Support
+
+- Email: [your.email@example.com]
+- GitHub Issues: [repo issues link]
+
+---
+
+## Development Documentation
+
+### Architecture Overview
+
+Describe the software architecture (diagrams encouraged).
+
+---
+
+### Project Structure
+
+```text
+root/
+├── src/
+├── tests/
+├── docs/
+├── ...
+```
+
+---
+
+### Development Setup
+
+Instructions for setting up the development environment.
+
+---
+
+### Coding Guidelines
+
+- Naming conventions
+- Code style
+- Linting/formatting
+
+---
+
+### Key Modules & Components
+
+| Module        | Description                 |
+|---------------|----------------------------|
+| module1.py    | Handles X                  |
+| module2.py    | Responsible for Y          |
+
+---
+
+### Database Schema
+
+Include diagrams or table definitions.
+
+---
+
+### API Endpoints
+
+| Method | Endpoint    | Description         |
+|--------|-------------|--------------------|
+| GET    | /api/item   | Get all items      |
+| POST   | /api/item   | Create an item     |
+
+---
+
+### Testing
+
+Describe test strategy, frameworks, and how to run tests.
+
+---
+
+### Deployment
+
+Instructions for deploying the application.
+
+---
+
+### Changelog
+
+| Version | Date       | Changes             |
+|---------|------------|---------------------|
+| 1.0.0   | YYYY-MM-DD | Initial release     |
+
+---
+
+### Contributing
+
+Guidelines for contributors.
+
+---
+
+### License
+
+Specify the project license.
+
+---
+
+> _“Documentation is a love letter that you write to your future self.”_  
+> — Damian Conway
