+++ config.py
@@ -1,0 +1,4 @@
+# This file contains configuration values and constants
+
+DB_PATH = "rechnungsverwaltung.db"
+UI_PATH = "./Qt/main.ui"
